+++ src/index.ts
@@ -14,15 +14,22 @@ export {
   merge,
   parse_csv,
   parse_json,
+  parse_table,
+  parse_tsv,
   pivot_table,
   read_csv,
   read_csv_sync,
   read_json,
   read_json_sync,
+  read_table,
+  read_table_sync,
+  read_tsv,
+  read_tsv_sync,
   to_csv,
   type ConcatOptions,
   type ReadCSVOptions,
   type ReadJSONOptions,
+  type ReadTableOptions,
 } from "./io";
 export { Series, type SeriesDType, type SeriesOptions } from "./series";
 export type {